Sort out the stacking order of menus and controls
Add the z-index setting to menus, menu controls and UI switches. Set such that
the menus stack in a reasonable order, even on small screens.

The controllers for the menus, notably the handles for the two main screen area
menus, need to be available as much as possible.

The settings controls take instant effect, including changing the styling of the
menus, so that panel takes priority over all others. The site navigation menu
recieves priority over the other menus. The on-page navigation will overlap the
misc. right-side menu if the screen is small enough, though that's well into the
low end of the 300-400 pixel range, less than 320px. At 320px the two handles
for the pull out menus will overlap, which was dealt with earlier by sliding the
right-side handle down.

For accessability reasons, the default color scheme was moved to the bottom of
color scheme list. On a 320px device in landscape mode only two schemes are
visible and the list isn't scrollable as it's tied to the footer which is fixed
to the bottom. Users in that mode will have to accept that limitation, or else
rotate the device into portrait mode.

Somewhere in the css file split the animation for the settings menu was lost, it
is restored.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>