Add styling to the title bar
Finally found a setting for the font sizes which seems to work well across
multiple devices, viewport sizes and user configured base font size. The
switch between layouts with displayed font being as large as possibe without
causing proplems when the user selected large font sizes in their accessability
settings was a puzzle to solve.

Added a few tweeks to the menus as well.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>