Setup for testing of fonts
The main page area was set at a controlled size to handle experiments with the
styles. With content comming soon it is time to set it for "production" styles.

The margin allows it to clear the handles for the pull-out menus. The bottom
padding is set to clear the fixed footer.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>