Add content for visual testing of fonts
Added what will likely be a page in the "site". It has enough mixed content to
provide decent flow review as I play with the font switches.

The content showed a few things which were still needed, such as a nice font,
and other settings, for code formatting. Added some of the styling for page
content, and an old-style font. The EffectsEighty font is designed to look like
that from the dot-matrix printers of the eighties. Also comes out looking nice
as a substitue for the old CRT monitors.

More content styles will be needed, but the fonts can be tested now. For that I
changed the size range available, dropping default from 4 to 3, allowing more
room to scale up than down. 16px default shouldn't need to scale down much, but
many people probably need to scale up farther. The top-end will now be 48px, or
triple the default for those with higher default settings.

In the hero-box container a single period is enough to cause it to be non-empty,
cancelling the hiding of it for being empty. It is just a big yellow box, but
it's there for me to "see" as I experiment.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>