Set side menus to always open for wide screens
The on-page navigation menu appears when the screen is wide enough to handle it
and the content will still look good. Going even wider, and the right-side menu
also becomes permenant.

When a menu becomes permenant it's length becomes more dynamic. As a fold-out it
is fixed in length. For a short list the means whitespace at the end, for a long
list it means a scroll bar. After becomming always open the length's only limit
is the screen height. For short lists the length reduces to what's needed, and
the scroll bar is only needed if the list is quite long.

Reformatted the HTML slightly, reducing indent for the main content.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>