Fix radius on menus with scrollbar
Allowing for a scrollbar on the side menus cost me the rounded corners I like so
much. The scrollbar, under the browser's control, overlays the element it is set
to scroll for, and ignores the border radius of its container. It, however, is
"under" the container of the container. The solution was to place a wrapper div
around the scrollable div and apply the radius to the wrapper.

Simple, in theory. Not so much in a complex setup. Had to shuffle some of the
styling around, especially where I used child selectors. Added to that was the
complication of using the same classes in several places. In the end, not much
actual work, just a lot of attention to the details and moving some rules to new
selectors.

As a bonus I found an old rule left over from the use of <ul> for menus, which
was dropped many revisions ago, back at commit 905c387.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>