Convert nav items into links
Till now they have been placeholders rather than links. The project is to the
point where I'll have to start adding content, and the links will need to be
active to make the site work.

At the same time I took a tiny step into factoring the CSS into functional units
rather than semantic units. Given time, I'll refactor the lot of it, and drop
the use of id selectors for the most part as well.

I'm still debating on using a separate sheet, individual sheets, or internal
sheets for page content styles. It seems that the common elements for the site
should be in the site-wide sheet, and that which is subject to constant change
should be somewhere else. Pros and cons for each option, and it's probably more
of an executive decision than a performance choice.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>