Automate the menu creation
Looking at making several pages, each with their own copy of the main and right
navigation menus, which should all be exactly the same, was a daunting idea. The
project requires me to use JavaScript anyway, so why not 'use' it?

Placing an object in the JS file to define the menues for the top and right, and
extracting the content of the page to make the left side menu removes the work
needed to build, and maintain, the menues.

Changed the archive menu to right menu. Archive was too limiting, and misleading
about what is likely to be placed there.

I might increase the menu automation to another level, or two, higher in the
DOM tree. Having done so already would have made my earlier changes for the
lost corners even more elementary.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>