Switch paths to be root-based absolute
Investigating the operations of CodeSandbox, since that's the place for our
final project to be "hosted", the sandbox does use the root folder the same
way a hosted website would. My development environment wasn't setup to work
that way, but can and will be now. That allows me to use the root folder as
the anchor point for all of my assets; JS, CSS, and images.

Adding in the CSS files changed for, but not staged in, the last commit.

Made fast copies of the files named in the menus so that the links are not dead.
Probably fill them, and others with Ispum Loram for now. Content creation isn't
what the class is about, after all.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>