Add previous coursework
All of the coursework leading up to now is planned to be made available through
the site. Have to have them here to serve them with the pages of the site.

As the site has been styled and themed, which will not match any of the work
done on the prior work, I'm thinking I'll use an iframe in a modal to "show"
the completed projects.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>