Add the imported countryCityStateJson package
Populating the country and state/province drop downs takes much more information
than I have the time and energy to acquire. I found a, presumably, good version
of the information, along with the Node.js implementation to us it at
    https://github.com/khkwan0/countryCityStateJson
With some modification to the JS to make it work in Vanilla JS I've added it to
my project.

The only file actully used from the package is the compiled database in the lib
directory. The whole package is included here, in the event I ever switch to the
Node.js system for this site.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>