Add the beginning JS for the contact page
So far the only significant code is used to populate the drop downs for country
and state/province.

Added my functions as methods to the object created by the imported package. It
just seemed best to operate within the promise section for the country list. It
is built on page load, and has to wait for the database to be loaded anyway. The
rebuilding of the state drop down is done each time the country is changed, so
that section need not be there. It works within it well anyway.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>