Restructore the directory layout
Spontaneous decision to declutter the root directory. Moved all the root-level
pages to the pages directory. Blog entries to the blogs directory.

Reviews, references and legals can each have their own directory. The portfolio
page might be a collection of external links with only screen shots stored on
the site itself. If so, they can be in an images directory without the need for
a portfolio directory.

Due to the change to dynamic menu creation the only work involved was adding the
changes to one JS file. I did discover that the logo wasn't using an absolute
link, however.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>