Adjust side menus for more space
Decided that the side menus ought to have a bit more room as the page gets
wider. 15% of page growth will go to the side menu already opened. The
center section will also be growing slightly slower as the page expands.

The contact page still has a bit of an issue with the flair images, but I'm
short on time at the moment. Later I'll try to iron out its operation as well.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>